Winsorize outliers based on z-score cutoff to next most extreme non-outlier value

The winsorZ function identifies outliers based on Z-score cutoff and replaces with the next most extreme non-outlier value. This involves z-scoring the variable and identifying/replacing any cases beyond the z-score threshold. The winsorZ_find function is an optional companion to flag any Z-score outliers to tally as needed.

winsorZ(x, zbound = 3)

Arguments

x
The input variable to Winsorize.

zbound
The Z-score cutoff (default=3, i.e. outliers are Z>3 | Z<-3).

Value

Output Winsorized variable

Examples

winsorZ(psydat$iq)

#>    [1]  93.83228 112.68802 106.11751  82.57565 103.96754        NA 112.43840
#>    [8]  84.98447 109.72077 102.00089 107.82870  94.24800 101.16819  65.79990
#>   [15] 121.33117 128.06143        NA 120.94659  99.31655 102.56419 103.13269
#>   [22]  75.09702  95.11244  91.19662 118.83513  98.69861  99.32034  85.26256
#>   [29]  77.90569 147.97605 133.67262 117.09030  85.56059  88.84899 119.96006
#>   [36]  85.13261  86.12356 121.29356  77.58612 108.93702 116.54272 131.46208
#>   [43] 111.66727 120.16203 113.45050 119.62288  88.63850 111.02213        NA
#>   [50] 101.80584 104.39425  97.83129 148.01383 129.87105  93.48121  96.92514
#>   [57] 126.67364 112.24818 131.96517  77.36263 116.16270 119.72661 108.97427
#>   [64] 105.45707  66.10560 111.54051  89.76036  89.79048  75.74993  91.30492
#>   [71]  89.64806 125.37786 144.43735 112.12403  79.67732  83.45306 102.78938
#>   [78] 112.98459 110.13161  71.32237  92.00815 108.86511 105.60069 106.27559
#>   [85]  95.44980 132.15718  90.66665  79.58028 119.84790  96.66050 101.86165
#>   [92]  81.62879  85.55559 120.58878  91.85833  74.67916  90.40886  80.17911
#>   [99]  91.25347 110.71767  92.68792  99.46792 116.42699  94.99791 112.65734
#>  [106]  93.33950 105.52749  89.67275  88.66027 100.75326  99.50824 126.37813
#>  [113]  98.36886  85.07169 100.47924  91.49633  96.47743  98.04574  82.63620
#>  [120]  83.52121 115.59302 102.97090 101.84950 111.12004  92.02469  80.08304
#>  [127] 124.49964 101.77897 122.00568 109.38164 104.92421 129.46809 100.32707
#>  [134]  98.71061 126.45942  72.89456  87.28485  93.64661 133.41390 110.51106
#>  [141] 114.75124  82.90347 113.17681  98.04431  90.21237 129.82166 102.39084
#>  [148] 130.03418  85.38217 123.97718        NA  82.22159 149.52278  79.60283
#>  [155] 121.81614  50.15526 109.09227  90.32188 100.91455 110.94856  84.32234
#>  [162] 101.38960  99.08970 118.92716 111.72320 111.52997 101.59428  83.17193
#>  [169] 121.83961 107.84881 110.03924 107.54544  95.47074 115.25286  93.81206
#>  [176]  75.22946 100.04458        NA  95.58525 121.30906  98.52900 127.85075
#>  [183]  72.45373  90.98642 127.40799 109.16020 100.46069  83.07221 142.37489
#>  [190] 111.16505 110.47812        NA  90.23753        NA 124.14114 106.28460
#>  [197] 118.56811  76.11961 114.55472  78.03294  69.82488 128.44262 126.59001
#>  [204]  95.09979  96.65467 112.70062 104.86545 113.39999 104.30055 120.20451
#>  [211] 125.72216 103.24042  89.79973 104.68447 102.30558 102.92480 121.30882
#>  [218]  81.21123 127.21928 102.84593 113.07585 114.26039  82.15454  90.44004
#>  [225] 105.28172  81.83072 148.45001 138.35348 102.95841 109.58794  83.30955
#>  [232]        NA 111.23666  85.26699  90.52241 120.36593  98.99688  83.12484
#>  [239] 102.48376  67.79238  89.54452 112.89160 104.38049  94.59043  91.03488
#>  [246]  88.10515  84.17265  84.88593 120.46759 107.30285  73.46040  94.76857
#>  [253] 102.48807  79.30056  92.64726  90.23512 113.76235 104.44152 136.58553
#>  [260]  83.01034  72.80975  85.72814 108.13866  95.43319  94.82922  70.31473
#>  [267] 100.10065 114.84287 129.84331 108.48397  79.86129 120.22680  65.94827
#>  [274] 117.51886 102.65406  94.19908 108.74283  92.21676 139.33824 107.12594
#>  [281] 117.13045 113.82782  91.50885  77.80407        NA 108.21784  93.83887
#>  [288] 104.58382 142.52188 111.72614 126.27361  95.77577  93.79237 104.73141
#>  [295]  83.45041 107.26041 105.52033  75.76365 100.68543 106.35127 118.33784
#>  [302]  88.52078 111.61773 104.43404 109.86807 115.02161  78.38297  96.48339
#>  [309]  96.22615 106.46892 128.88644 123.08892  97.40423 100.47120 100.02640
#>  [316] 107.06511 103.86386 100.49375 126.29119  90.87513 104.14333 110.05974
#>  [323] 107.73991  82.34723  64.88359  86.93736  89.99352 101.49047  82.98481
#>  [330] 111.19818 118.18837  94.83440 120.71834 100.36548  98.66100        NA
#>  [337] 125.24477  97.05519 111.80335  94.22725 105.68228  76.53873  84.46584
#>  [344]  79.40343 102.01760 104.92746 101.00085 127.34086 104.63064  72.00372
#>  [351]  76.68754  68.00840 102.43669 116.24726  99.58570 140.49713 117.00163
#>  [358]  76.46627 145.26168 102.66307  88.85506 112.77419 116.36395 101.95612
#>  [365] 119.86329 140.07574 118.82892  93.67413  90.39221 104.14761 108.62487
#>  [372] 119.42499 112.63181 101.85870 112.41028  77.78464  79.28116 111.55982
#>  [379]  84.24223 111.89636 106.76049 100.93298  89.50566 106.92587 101.45924
#>  [386] 114.81745  99.79988 143.31503 113.73850 138.88941 100.78805 109.35671
#>  [393] 107.81766  73.21392  85.18986  65.33477 102.44323 116.91728 118.44811
#>  [400]  80.25918 116.63301 104.54019  87.78987  92.44125 109.42658 113.27099
#>  [407]  80.62082 110.59745        NA  83.44222 101.26607 113.77545 114.47837
#>  [414] 151.95282  85.69496 106.62307 121.42092 115.24731  98.09602 130.60070
#>  [421] 116.94279 101.19407 123.06786 107.21265  94.24608  77.12840 102.01065
#>  [428]  94.48165 103.49908 110.29107 114.41729        NA 127.87850 130.25569
#>  [435] 104.81304  95.41706  93.01583 140.05386  91.77482 100.30781 112.27969
#>  [442] 113.54755  86.09634  84.63980 102.52176  66.68521  98.94638 112.74622
#>  [449] 103.35428 117.49256 125.13185 108.75579 124.19881  77.55519 121.91352
#>  [456] 105.90817 102.46196  92.55656  77.64579 104.02762  91.14183 123.01554
#>  [463] 117.98859  96.92604  67.24854 124.82553 101.67806 112.45925        NA
#>  [470] 105.83682 114.18714  99.88288  96.56357 117.90442  80.19385  71.26340
#>  [477]  59.45710  91.99778  91.08088  82.76447  78.02502 106.94039 138.11558
#>  [484]  94.29296 116.26338 119.31706 100.56532  79.87917  98.57569  97.84712
#>  [491]  69.51702  80.30071  88.24708 145.58603  92.54786 118.16008  89.75926
#>  [498] 103.43221  96.16064 112.46027  88.20953  99.27020  94.15569  83.45542
#>  [505]  74.80722 121.10028        NA 102.25720 156.11932 108.54983 130.56407
#>  [512]  99.94773 120.54975 103.54448  95.33630        NA  97.45015 105.75562
#>  [519] 114.39929 108.86404  94.86913 114.77880 106.92381  85.65444  97.64873
#>  [526] 103.09441 116.50118 130.71452 126.63137  77.35727  83.87394 113.66289
#>  [533] 115.48144  92.93727 101.15415 102.87205  84.96968 107.08372 113.62503
#>  [540]  77.36580 116.30492  86.25881 104.25014  83.92176 101.78215  84.29672
#>  [547] 104.12837 102.77553 113.59183 112.10346 102.46733  84.54325 100.37142
#>  [554] 125.74570 108.92614 132.06097  95.60623 112.17987  97.48013  96.82616
#>  [561] 117.86648 102.24813 156.11932 121.89859 111.80622        NA  92.51357
#>  [568] 109.71569  81.39734 102.68381 100.39705  91.06652 112.46293  86.70125
#>  [575] 101.54761        NA  83.22300  81.53571        NA 110.57510 102.69510
#>  [582]  81.31967 126.72328  65.10386  89.89749  95.26914 128.77402  97.90562
#>  [589] 119.97806  93.93612  81.37287  89.60425        NA 129.98005  86.72246
#>  [596] 113.25368 112.07210 110.64343  94.41640 110.19478  98.53073 102.06922
#>  [603] 111.11547 106.81064  99.83636  87.94785  92.71070  82.54563  82.34842
#>  [610]  96.56793  94.60107  92.13193 119.66805 107.66492 118.06232 115.70433
#>  [617]  90.41811  90.24873  96.01093 100.67706  72.10595  92.96797 111.55053
#>  [624]  75.23611  80.58179 106.87106  84.04066  70.59569 119.20901  78.06134
#>  [631]  87.51366  99.95829 113.43522 121.85312  78.16464 116.76835 115.31708
#>  [638] 135.43305 129.35233 112.10612 106.28115 117.51176 119.15623 109.34158
#>  [645]  78.56968  84.00552 131.97113  74.85827 122.07172 112.48462  91.18679
#>  [652]  70.64130  93.68338 101.76688  91.19509  96.24813  89.30254  68.13336
#>  [659] 125.08575  95.52732  87.30261  99.78592 119.76620  88.03863  79.83555
#>  [666]  72.64333 107.47919 115.03443  86.36076 113.08313 123.39112  91.75911
#>  [673]  78.10123  72.23790 117.68751  91.88840  78.18220 141.33503        NA
#>  [680] 102.86817  78.03841  93.78726 106.20320 118.87789 114.26603 107.51318
#>  [687] 114.94192 123.66976 116.81471  97.88588 104.21496 105.70631 102.80319
#>  [694] 110.47101  90.93592 111.91550  87.62613 116.35673        NA 104.50097
#>  [701]  98.63290 106.11178 123.84395  89.01710  89.65978  91.98903  89.57789
#>  [708]  91.99023 122.90727 100.90405 120.04355 123.02503        NA  89.93316
#>  [715] 132.37785 119.63569  87.98856 103.71023  85.66844 121.78801  96.70289
#>  [722] 103.86545  84.29271 108.84636  90.42721 100.78925  81.44827 107.85019
#>  [729] 109.84869  90.36456  77.74056 104.30153 104.32588  74.43073  80.36561
#>  [736]  76.98214 106.96607 110.67162  99.04811 110.43475  88.02531  95.30663
#>  [743]  94.71710  72.58595  99.44863  73.80015 108.26119 103.65948  88.34191
#>  [750]  66.72763 113.15649  99.68802 122.84583 110.42012 105.58653  99.27441
#>  [757]  62.20654 124.02486 105.96096  74.28921  85.64990  91.04417 156.11932
#>  [764] 124.61901  94.49176  78.26397  99.64379  66.11195  78.36850 105.73386
#>  [771] 106.06434 103.68987 114.90803 115.70212  82.64791 111.36369 114.04764
#>  [778] 112.80309 134.51807 140.84410 107.13148  82.13119  89.49511 119.45493
#>  [785] 103.81357 115.17862 117.46303 100.70601  89.98454 116.71043  96.49347
#>  [792]  99.31274  97.16255 150.30420 111.90070 122.82553 111.67263 115.62123
#>  [799] 106.93744  89.96936 103.54924 111.17832  91.25830  93.93155  72.77662
#>  [806] 107.86428  95.77609 137.52677 140.93103  80.22276  96.33515  72.67807
#>  [813] 102.34485  60.74098 109.29436 111.65475  90.55841 104.05541 115.30740
#>  [820]  99.28138 104.79916 118.65996  94.61509  80.42603 122.43775  72.17324
#>  [827] 114.69133 104.05847  66.15348  87.04426  91.65180  99.20192 111.43259
#>  [834]  94.11257 120.87917  84.69644  85.06698 130.40575 110.13158 101.43239
#>  [841]  97.37508  93.60192        NA 135.65343 118.90984 110.24815 124.79628
#>  [848] 113.81419 106.72332  94.80510  71.07237  81.73260  74.20493 115.50920
#>  [855] 100.04372  93.07552 109.23256  64.60033  90.92300 110.89289 103.38133
#>  [862] 132.78193 109.67601  85.54169 106.06511  85.72384 118.52997  89.68161
#>  [869]  88.18830  67.99915 127.86639  83.35205  78.00624 112.96771 140.19736
#>  [876] 143.85204 122.61594 112.86236 133.05175  83.38429 101.34510  83.95736
#>  [883]  89.11675  95.13383  97.21707  85.62241 122.67251 127.94905 111.05761
#>  [890] 103.24193 100.06708 106.70110 130.60696  97.00769 105.36706  80.53210
#>  [897]  86.28552  84.36419 117.19135 124.14583 112.08913  97.24859 100.44071
#>  [904]        NA 142.98170  86.48340  99.64217  78.87779 156.11932  95.60828
#>  [911] 130.10493  92.84180  86.97361  89.53391  78.22045  74.25249  74.65079
#>  [918] 110.27572 128.90499  99.34116  96.13674 103.30348  86.27168 106.95548
#>  [925] 113.11812  99.86588  97.05767  86.86917 126.88542 123.70767  92.71309
#>  [932] 120.10141 119.52864  89.72849  85.07340 108.91783 104.59831 128.39144
#>  [939]  91.70361 106.51584 119.68743  97.06056  90.18468 127.28296 105.36223
#>  [946] 106.29019  99.29472        NA  93.41412 104.59694 114.70420  72.46266
#>  [953]  62.85491  81.44541 131.55296  99.36234 119.64501 130.36915  99.99948
#>  [960]  95.17776  93.97166 117.85989 103.30341  78.45337  80.14520  81.63814
#>  [967]  90.81994  83.73286  96.89699  79.56631  77.63782  58.12543  92.76005
#>  [974] 118.69229  95.62343 108.70905 106.35359 103.49547 127.05559 106.57386
#>  [981]  85.84504 119.12112        NA 112.57523 114.78396  93.79565  89.22705
#>  [988] 111.89458 148.88985  91.49753 119.18822 112.97886  77.54702 102.23980
#>  [995] 106.68501 101.82487  99.01983  75.40306  88.10180 113.49464 121.65170
#> [1002] 156.11932  90.74233 123.21508  87.31172 122.19103 124.67818  93.84515
#> [1009]  81.09299  81.36062 121.98918 108.17035 111.41980  97.09627 103.28013
#> [1016] 108.74840        NA  97.83828  84.32722  91.05704  93.37573 111.47026
#> [1023]  71.49631 120.92622 116.68801 133.12338 121.93352  72.41006 117.43847
#> [1030] 103.99941 109.57697  92.75551 116.44371 135.80996 126.75774 136.44292
#> [1037] 109.73057 110.53055        NA 128.32910 113.96033  69.01524  76.50648
#> [1044] 109.67423 110.33984  93.28339  92.11099 100.75649  97.36911 119.25964
#> [1051] 100.71708 113.78079 129.65226  93.47223  71.81652  99.79891 137.96479
#> [1058] 111.96208 126.46649 101.36804 139.04501 111.55025 114.19275 114.98869
#> [1065] 126.79223        NA  98.42188 138.78874 111.08784  81.31437 128.56572
#> [1072] 113.44102  83.88640 110.38143        NA  96.49724  87.04243 122.61503
#> [1079] 140.63194 130.85579 110.99299 107.09089 100.07612 113.25448 102.04789
#> [1086] 119.51847  84.70255  92.90023 119.74532 119.93391  69.01253 100.33212
#> [1093] 113.81166 104.03485  74.33540  85.79473 137.04017 106.75976  96.21196
#> [1100] 114.26823 102.97980  83.10550 109.21359 120.83226 106.44629  89.76213
#> [1107] 127.59770        NA  75.92489  89.04250  99.49008 107.27043  99.56003
#> [1114] 101.36484  73.04833 100.42771 103.33044  89.20976 106.42483 103.19405
#> [1121] 108.99492  99.66365 110.08043 130.26892  98.02472 119.17712 103.12997
#> [1128] 103.02179 106.54420  89.24513  91.32144  97.66684 107.45344 122.68687
#> [1135] 100.38633 108.56607 128.88392 123.16571 112.03640 106.84816 124.64605
#> [1142]  96.84758  98.78559  99.44104 104.96877  97.81333  97.32597 113.23737
#> [1149]  78.65068  99.10356 137.50021 116.12966  98.47461 115.21288 102.68502
#> [1156]  89.33228 110.95279  95.99139 104.87896 145.18179  93.25393  92.60118
#> [1163]  66.47308 108.41002  87.88049  61.55268        NA 156.11932 107.43584
#> [1170]  73.50219 110.15173 114.04371 100.52931  99.65120  97.72000  99.25940
#> [1177] 114.06588 104.13127  69.25153  75.11007 119.67675 109.98219  91.63695
#> [1184] 118.42511  87.30426  92.04464  80.15521  95.73935 105.63835 128.05677
#> [1191] 102.21587 106.73541  99.54778  84.20371  87.78621  81.41722  82.78118
#> [1198]  90.87002 100.58541 117.74498 156.11932  87.76292 117.44268  99.42498
#> [1205] 105.82968 104.80979 147.58069  86.18069 111.69104  76.95038 112.22604
#> [1212] 122.25109 103.83039  79.88355  77.67091  77.51395 120.79501  84.72709
#> [1219]  94.55586 111.05669  79.86266  90.34025  93.06179  85.00005  91.33889
#> [1226]  87.97341 100.50785 100.49127  82.40587 109.06450  91.52739        NA
#> [1233]  93.70088  93.67248 105.30942  77.83091  86.29471 104.45570 103.69788
#> [1240] 115.92816 107.16211  94.92444  98.58795 102.08042 107.47594 103.66263
#> [1247] 104.92769  85.93170  95.10906 109.85359  75.33102 111.38575 105.87721
#> [1254] 117.64356 107.51270  98.60616  85.08805  94.49022 101.44285  95.87911
#> [1261]  85.15855  86.98613  92.07825  99.85727 114.17843 106.03928 135.59854
#> [1268] 106.27675  97.66258  66.90156 114.08770 103.46169  86.86859  71.26625
#> [1275]  83.43568 125.56898  92.77513  95.64448 109.29560  98.24246  84.54503
#> [1282]  99.39822 103.23241 100.54626 107.87738  97.11237  88.72274 107.49854
#> [1289]  82.39039  80.56339  87.14395 115.51459  96.36379  96.24323  77.13400
#> [1296] 113.34493 100.23090 128.00318  83.61198 110.78141  85.74094 123.16767
#> [1303]  86.91226 108.71729  85.19678 106.03960  82.50363  90.26754  99.95278
#> [1310] 118.03201  97.51849  96.93912  70.43535        NA        NA  93.32997
#> [1317]        NA  78.62687 128.92895 119.37688 116.07438  93.51287  93.98273
#> [1324] 101.55616  78.49652 107.99965  82.75218  89.03718 120.34173        NA
#> [1331]  70.03057  84.91780 107.79848 114.93714  88.04591 111.23774 135.94260
#> [1338] 111.76744  68.31509 103.69276  74.64519 105.53407  83.10455 113.75842
#> [1345] 107.56972 122.28815 115.69738  78.91743  65.07201 109.35336 109.79649
#> [1352] 102.93410 109.82011 151.31432  86.78238 109.38561  71.92861 143.30574
#> [1359]  95.48715 107.86270 134.56015 125.21975  81.94606  68.47050  74.60976
#> [1366] 120.25062 113.87796 101.72002  85.96520 120.87913 110.15029 118.40815
#> [1373] 122.21053  95.21304  90.13372  85.72207  94.90467 119.81935  89.85913
#> [1380] 112.02081  92.67920  99.37070  88.00582  87.47771 156.11932 118.91114
#> [1387]  90.03121 134.81421 104.43641  98.47255 109.51881  92.06761  80.12481
#> [1394] 103.16246 110.90524 106.85035  91.63824  97.21924  95.52354  88.17483
#> [1401]  79.76704 124.51006  89.04053 104.71264  94.46662 138.61166  75.72310
#> [1408]  58.33635  99.35342 134.55064 110.20590        NA  85.65238 108.32365
#> [1415] 108.98267  80.44252 121.49243  90.23953 102.55163 118.18259  94.24799
#> [1422]        NA  83.98722  91.97671  74.08123  84.73338 102.02263  82.79661
#> [1429]  92.39942 109.33542  73.76474  95.75765  88.08411 101.04090 129.97102
#> [1436]  99.11919  96.66612  83.54372  86.63223 128.09332  89.45786  96.95740
#> [1443]  76.18425  84.65547  79.09053  97.87731 112.02023  80.38542  90.18477
#> [1450]        NA  86.40029 108.87888        NA 101.56153  89.99190  74.82382
#> [1457] 109.95697 100.27328 103.42038 108.82663        NA  64.33090 101.21458
#> [1464] 102.24084  94.66066  78.15403 103.83676 124.85045 115.07187  96.68498
#> [1471]  96.91678  99.34967  67.07347        NA  78.81012 105.80887 101.51873
#> [1478] 111.32752 101.99411  98.33770  75.16672  84.76168 100.12868        NA
#> [1485] 104.63135  86.33809  98.15684  80.35380 142.66768  93.40471 126.51777
#> [1492] 124.59472        NA 103.38369  93.66014 110.95150  69.22883  97.59130
#> [1499]  94.03507 101.06470 105.10013 100.94643 125.49519        NA 132.52427
#> [1506]  97.55146  92.82272 108.57479 114.41895  69.25039  96.28345 100.92121
#> [1513]  95.60826 114.70484  80.18024 104.81314  93.79603 141.08371 123.13971
#> [1520]  92.16734 106.68228 120.37453 107.74534  86.85637 117.51976  91.04322
#> [1527] 124.44986 124.60340        NA 139.06007 125.55685 106.16419  81.54872
#> [1534]  89.31496 126.36544 133.50321 103.63943  98.33632 123.63071 114.02038
#> [1541] 111.74706  97.15933  80.77192  91.13169 113.03940 128.01154 103.44638
#> [1548] 101.28218  83.19146 120.34710  93.17602  95.14987 111.14346 107.48858
#> [1555]  76.43997  88.96274  59.33211  87.43341 108.40758  99.90051 114.25775
#> [1562]  84.22704  90.19903 101.25919  99.82763  74.09705 113.28964 117.02187
#> [1569] 118.49354 101.74745  93.29189  96.18272  78.93636  89.67790 112.59315
#> [1576] 117.16126 100.91470 104.99887 102.39245  91.92053  81.57011  95.65008
#> [1583]  94.83123 117.29104  72.71900  86.70674  95.81123  99.57971  82.52904
#> [1590] 107.68461  92.80510  76.22008  93.02643  97.44186 119.99923 112.59903
#> [1597] 119.01048  85.27966 108.38226  87.13057 106.26597  78.83461 115.17615
#> [1604]  99.29849 122.44010 110.51891  76.15139  95.69812  68.20659  95.24476
#> [1611]  95.66683 118.27221 111.52030  75.50186 110.14511 109.25048 110.98266
#> [1618] 102.46063 122.69539 111.29135 117.78743  70.96518 122.39869 116.29299
#> [1625] 116.66983 101.24009 111.66242  91.40206 108.16613 156.11932  97.08920
#> [1632] 100.55864        NA 107.29729 138.55963 156.11932 105.60255  52.74708
#> [1639] 106.11534 114.24327 107.85035 111.12282  94.30993 109.69117 140.37801
#> [1646] 118.94867 103.23225 124.07003 109.93699 103.05940 113.65853 109.30679
#> [1653]  60.18821 106.44874  99.08928 115.27588  76.39239  82.74149        NA
#> [1660]  95.56816  87.68606 110.90002 108.77314 124.15060 101.93165 113.87838
#> [1667] 111.76163  91.63451 137.38297 123.75283  96.11539 115.73121 134.49387
#> [1674] 100.52058  89.44425  99.85053 119.58549  81.69782  93.76666 100.68082
#> [1681]  97.24757 108.05351 107.13347 108.73612  88.87075 119.15300 102.67872
#> [1688] 133.47277  87.00165 126.64915  80.69885  81.60104 138.52344 119.67600
#> [1695] 100.30174  86.04988 113.18866 110.39543 111.59393  78.46024 134.51343
#> [1702]  82.53143  88.89758  91.79795        NA  78.93162  80.18600  87.20071
#> [1709]  64.34463  91.57780  85.12213  58.59312 109.02232 101.02661 112.17773
#> [1716] 113.13484  94.74464  94.72939 105.76462        NA  85.62127 118.77275
#> [1723]  77.32623  75.49475        NA  91.52053 136.50684 110.93723 109.82082
#> [1730] 109.15939 122.84025  88.80293  94.35747 122.94813 110.44645  92.30357
#> [1737]  86.30743 109.32227 115.45089  72.25852 135.79528 117.36181  76.55360
#> [1744]  62.92264  97.45713  55.15116 113.46305        NA  99.25036  93.75744
#> [1751] 107.32381 111.74612 105.52129 104.32111 108.95291  80.75474  79.15205
#> [1758] 105.41605  72.41044 106.92163  98.64777        NA  88.30216  80.70323
#> [1765]  87.57626        NA        NA 138.26623 106.76802 101.52735 124.08405
#> [1772] 102.77619 102.99895  82.45972 100.96786  92.10373 103.68480 110.18180
#> [1779]  91.54937 117.69453 101.95644 118.62911 106.24413  91.07526 118.07929
#> [1786]  97.37180 104.61603 114.43442 113.44194 101.41302  74.95951  95.27593
#> [1793] 132.42966  90.20622  94.76733 131.51565  95.25433  94.43482 139.33826
#> [1800] 107.76581        NA 116.67727 107.85589  81.58460  79.40238 103.42131
#> [1807]  82.43049 118.59917 114.37027  80.23371 118.15649 116.62848 113.76126
#> [1814]  97.80665  97.17441 103.44236  98.54288  86.64463 109.51632 128.56774
#> [1821]  63.04097  81.58369 102.24529 127.10736 112.27223  82.34521 102.03400
#> [1828]  98.35328 101.94591        NA  85.73632 135.60236  93.43564  69.70006
#> [1835]  91.53852 107.65097 141.20112  76.29232  91.72272  82.06497  92.80924
#> [1842] 100.62722  92.26582  91.71454  69.02751 116.43726 107.43750 117.62997
#> [1849]  89.09479 106.98657  95.89466  88.16936  92.76170 128.95622        NA
#> [1856] 129.20576 123.36692 113.93832 118.27554        NA 113.46043  81.67435
#> [1863] 103.27524 114.54700 120.99166  92.07187  98.75666 111.17105  82.64757
#> [1870] 100.75860  91.85334 104.81936 135.05057 103.02422  96.14114  95.41574
#> [1877] 111.47428 112.62669 124.31407 126.42647 112.63399 105.64271 136.66675
#> [1884]  79.99824  90.59379 112.70825 101.25045  93.86656 119.71380  94.90082
#> [1891] 113.31411  77.46923  87.80903        NA 107.90199 116.28393 119.57921
#> [1898] 115.48943 133.06015 114.78072  71.77798 101.38541  80.83294  78.11030
#> [1905] 115.93531 103.84547 120.74645  76.65640  95.68025 123.38669  85.02684
#> [1912] 107.36726 105.29635 117.89102 124.58703 138.21402 107.70824  90.67284
#> [1919] 116.36957  94.18994  54.50559 115.54010  82.94561 114.31045  92.93468
#> [1926] 100.58136  72.90178 106.29481 103.46819 112.22187  95.40602 116.87610
#> [1933]  76.11791 111.88005 121.98060 116.15947  67.48213 110.23634 106.64378
#> [1940]  98.45694 115.02767 110.78574 123.38860  98.33947 136.38411 114.96116
#> [1947] 114.72821  95.75631  90.21978 103.85271  85.11601 128.03830  70.81453
#> [1954] 130.05213 124.90082 114.25302  51.53648  96.48675 130.37140  78.61466
#> [1961]  48.74417 129.78688 100.69060 143.26548  93.70363  90.21938  91.33576
#> [1968] 103.71527 152.53888        NA 109.90144 100.50578 106.17849 108.24041
#> [1975]  88.09990 113.07445 103.80389 122.17108 106.97122 114.50394  79.29790
#> [1982] 105.44922 103.86642 122.50243  75.97465  79.38572  97.85368 138.06264
#> [1989] 119.30734  92.01229  91.28856 101.36988  76.05129        NA 101.38120
#> [1996]  89.29261  86.39192 111.98213 114.89879  90.40341 114.39341  88.95830
#> [2003]  98.95869  89.59708  87.42416  93.22190 121.54038        NA 121.45296
#> [2010]  95.24425 130.99747        NA 111.70843  81.60419 105.88741 117.76297
#> [2017]  99.83190 118.42421 102.23295        NA  92.78363 103.78995  86.84799
#> [2024] 112.78387  80.80390 147.28250  95.71034 104.20342 124.55005 105.07740
#> [2031] 101.34286 130.85143 111.67428 128.55419  92.68781  93.12018  86.39106
#> [2038] 110.94352  94.23932 116.49337 100.38671 112.75673 125.83014  97.38449
#> [2045] 102.34665 113.27800  75.43690 109.66339 103.63420  83.44111 130.87083
#> [2052] 130.07820  97.69257  92.26038 123.42155 131.20790  68.62195 125.51068
#> [2059] 137.30078  96.21808 129.48093 119.69934  90.87225 119.19000 115.31026
#> [2066] 109.87082  82.52993  88.05142 110.97817 106.95994 104.24945 104.49911
#> [2073]  90.57966        NA 107.78722  87.78924        NA 112.37345  82.71078
#> [2080] 109.61489 105.89413  69.95262 119.90517 105.19620  88.65882  82.34209
#> [2087] 116.91788 101.70527  80.90767 110.51574  93.94096  75.24947  99.18061
#> [2094]  78.60730  84.33655  90.67594 106.81559  88.38238 102.53568 100.48529
#> [2101]        NA 103.33540  98.60964  99.41966  90.14221 113.23195 120.32213
#> [2108] 142.62108 114.30147 101.78447 111.67117 107.95692  78.03367  80.36900
#> [2115]  83.57441 104.74476 104.62088  88.18902 102.64890 106.02120 108.24139
#> [2122] 101.06230  84.64189 109.41620  93.74955  90.57288  97.63438 137.72942
#> [2129]  86.74393 147.05519 106.83212 116.63057 119.33291 113.30130 132.76914
#> [2136]  96.91222 121.86224  95.96787 104.54919  91.50366  91.54122  79.01191
#> [2143]  80.81284  90.55713  87.49329 118.85273 111.84515 110.13540 119.27187
#> [2150] 134.75446  99.69905 105.13098 107.57160  93.83460  80.45858 101.34103
#> [2157] 101.02193 120.16914 121.32102  92.94093  94.18060        NA  96.33288
#> [2164] 108.99613 126.51327 133.64351 110.96055  92.44065  93.43963  77.00418
#> [2171]  78.22975  98.87572 111.17883 106.24519        NA  91.49946  89.46951
#> [2178] 113.00999 104.51756  99.08230        NA 109.62855  86.97321 118.03785
#> [2185]  70.48399 109.51063  79.78354 122.34747  85.39559 153.00135 104.39744
#> [2192] 135.66188        NA 110.45398 120.50931 109.75713 129.27192 115.04103
#> [2199] 116.52012  80.89475 124.47317 105.72171  87.08266  78.56410  99.78902
#> [2206]  96.85285 111.16641 106.90224  88.79055 101.91418  97.43886  86.85036
#> [2213]  73.46592        NA 130.80316  96.69193  94.02232 101.60502  69.70992
#> [2220]  99.70130  92.47133  90.26977  98.81736 139.06330 110.09765 127.67429
#> [2227] 106.65789 102.30540  88.07589  87.29520 106.73687 147.30977 112.31212
#> [2234]  89.67971  94.62521  88.38150  88.89506  67.55810 116.93135  81.73050
#> [2241]  94.43329 113.33274 102.38684  90.86233 112.77258  98.14863  97.24148
#> [2248]  86.95608 117.66920  96.91126  68.40144  82.11689  92.71193  60.87572
#> [2255]  95.86410 110.43996  98.29899  93.03413  67.79576 111.21336 115.97854
#> [2262]  97.40964  99.25200 106.25096  99.49820  91.77741  82.69989 113.90538
#> [2269]  81.92327 108.24188        NA 132.36247 121.08522  80.91498 107.81731
#> [2276]  96.40050  67.59068 120.91765 119.82398 125.01455  90.24565 116.88094
#> [2283] 120.15999 101.06760  96.35741 108.31182        NA 132.24978  67.06285
#> [2290]  74.81338 119.94283 106.70632  72.85693  75.18872 109.83121 101.28885
#> [2297] 125.32909 122.56426  91.98634  90.48584  74.21737  82.02084  86.62967
#> [2304] 106.82668 109.83314  81.65184  64.92332 115.13213  95.01076 154.91425
#> [2311]  98.14731  48.74417  88.92823 112.26009  98.03837 118.34558  66.95878
#> [2318]  95.36874 109.16480  98.53322 127.91364 109.65725  97.55899 105.81095
#> [2325]  96.61310 101.42044        NA  81.73748 129.59797  94.83003 100.48409
#> [2332] 117.77078 116.76028 109.04354 103.54483 104.20533  67.73448 101.29387
#> [2339] 104.67198        NA 126.25568        NA 118.97123        NA 103.29011
#> [2346]  64.36424  88.57401 104.43366  88.90674  67.83804  91.09862 102.00271
#> [2353] 103.77852  63.41004  95.82096  93.33408 119.49632  85.31560 113.44504
#> [2360] 104.29128  82.79675  95.73350 145.97747  93.00328  94.83247  92.57157
#> [2367] 122.56063 115.96224 105.76750  88.17437  82.44883 108.80723  90.28821
#> [2374]  82.48601 117.82675 142.12380  93.17771  96.14807  77.04207  98.80633
#> [2381] 107.47086 101.26637 127.94089 113.96632  96.66689  87.04159  75.59485
#> [2388] 105.28895 107.58631 103.58129 101.19564        NA 101.80979        NA
#> [2395] 133.50512  79.60107        NA 124.86854 117.35296  92.77158 101.43699
#> [2402] 113.07600 127.11704 100.21687 115.54839  90.94378 135.49510 100.99690
#> [2409] 105.28491  86.42118  66.43361 113.63133 101.36221  82.76224  95.81707
#> [2416] 108.90875  82.88296 118.13103 139.28826  96.06681 100.33052  84.97597
#> [2423] 141.36079  98.34275  74.81665  90.39058  71.45503 115.05062 114.90466
#> [2430]        NA 107.60803 113.92464  79.39665  83.53702 118.40062  74.71015
#> [2437] 102.73009 103.63756 125.91570  96.14565 118.28682  90.18669 151.98909
#> [2444] 123.05407  94.77989  75.48654 115.59261 133.81575 117.87127 108.90641
#> [2451]  89.51751 102.76426  98.32881 119.62652  81.12333        NA 119.10866
#> [2458] 109.16896 103.88480  86.89558 100.70204  71.77162 106.59629 123.78357
#> [2465] 126.77999 104.10687 138.19843  89.31662 109.91544  83.66794 102.73209
#> [2472]  84.30284 108.76649 101.97787 108.22755 108.30732 111.67980        NA
#> [2479] 104.51130  98.09896 100.06353  97.04457        NA 130.53733  92.95155
#> [2486] 112.20193 121.30439 136.22106 135.13092  76.51965 150.57409  96.68416
#> [2493] 114.02920 111.33302 140.40475  95.35459 104.43319  70.11790 114.21679
#> [2500]  87.10609 117.72283 102.48420        NA 129.78499 118.98173  59.90796
#> [2507]  94.34317 101.70931 110.65878  94.78597  88.60461 145.90567 122.80334
#> [2514]  85.70685  97.03198 100.30462 106.90116 100.34483  77.95252  70.90410
#> [2521] 101.58304 119.92898 105.24952  74.85186 109.81079  94.20477 110.27334
#> [2528]  92.94037  89.77669 130.72936  97.69237  86.61092 153.84851 131.54174
#> [2535] 108.05385 126.14956  79.16820  96.30627 116.51767 124.75427 104.68401
#> [2542] 104.26529 108.74068 120.81686  59.29544 100.04091 110.43773  91.18146
#> [2549] 112.94520  89.65800 102.31487 109.01368 126.17960  84.79913 103.31301
#> [2556]  98.16376 115.48970  89.11828 113.13406 152.02780 100.29777 111.59034
#> [2563] 124.70768 101.62805  83.39978  74.59196 104.79103  92.44017 107.24076
#> [2570]  90.27371  90.13133  93.26680 101.61617  70.99708  94.87187  81.33162
#> [2577] 124.76760  84.31316 137.00933 100.25792 115.38491  89.55593 111.81711
#> [2584]  74.02949  97.43543  88.52234        NA  71.97052  96.25725 108.26785
#> [2591] 107.00815  91.89412  76.08529 120.62670  96.51881  93.49006 120.40623
#> [2598] 113.38094  86.04416 125.70383 101.09674 105.01714  93.76789 132.04040
#> [2605] 111.91348  79.37705 129.74422 108.13942  77.81027  92.15262  76.01619
#> [2612]  81.64617  97.51452  77.69148  88.70858 111.51461  94.63330 108.39322
#> [2619]  73.44398  80.52694  90.43804 106.42419  90.97631  94.78859  93.25900
#> [2626] 112.00852  75.38894  90.49079        NA  93.14600 111.76149  85.57796
#> [2633]  70.86207  95.69926  95.87197  99.70747 156.11932  81.96818 113.97600
#> [2640]  86.00931 101.32687 156.11932  97.12748 135.14151 103.45577  91.83131
#> [2647]  92.65933  97.20711 119.44740 101.12906 116.08613  88.19582 130.81421
#> [2654]  81.86093 142.45611 156.11932  83.94412 104.85294  93.89862        NA
#> [2661] 107.99308 109.65977 130.72975 105.97250  76.46146  94.05066  88.82760
#> [2668]  96.90910 112.19321 111.79576 113.35912 103.86284  92.95579 106.11491
#> [2675]  86.84984  96.41179 122.08478 104.46957  76.37238 154.21733 114.50837
#> [2682] 110.52202        NA 137.40260  88.95429  73.87442  77.68841 110.96618
#> [2689]  95.76035  92.76227  92.18345 119.57045 101.18077  89.87429 128.12353
#> [2696]  89.13509 115.48466  86.11954 108.94418 134.47603 117.32291  70.23538
#> [2703] 123.53188 104.33835  90.85609 106.73616 115.50729        NA        NA
#> [2710]        NA  97.86064  94.77916  71.43501 126.71685  91.94670 130.77439
#> [2717] 144.12001  97.04444 113.72518  79.88760 100.71213 115.11855  99.79579
#> [2724] 116.91929  68.44929 110.74154  69.90568 119.32582        NA 107.86980
#> [2731] 107.70156  95.04554        NA  83.59839  82.06138 105.75902 100.28624
#> [2738] 120.05214 101.86957  89.97448 111.45088 108.34445  90.81626  88.60447
#> [2745] 123.89292  73.79719  89.53423 101.94252  76.33821 105.11083  67.85904
#> [2752] 123.90695 109.21961 106.77674 113.77973 152.31517 116.72488 116.62350
#> [2759]  96.65717 152.30159 116.55184  85.80649  76.42908  86.24796  91.10839
#> [2766] 155.30165  95.58818 110.20653 101.90405  91.48217  70.98319  77.21469
#> [2773] 110.70737 112.02256  98.41174 109.07841  97.16729 117.87779 102.77851
#> [2780]        NA  98.95339  96.24799  69.22769 104.55120  99.89902 139.95315
#> [2787] 124.96171 129.82302 148.79349  94.54808 123.39720 146.68058 109.56701
#> [2794] 132.63600 115.66189 102.38717 119.34881 104.73586  95.29269  99.44166
#> [2801]  90.19293 148.70330 101.30528 111.28160  78.93075 115.72280 110.80103
#> [2808]  93.06227 110.85284 103.98486 104.77051        NA 103.22332  72.46926
#> [2815] 101.99606 106.06161 135.76997 122.00031 121.86987  97.12712  63.79696
#> [2822] 113.28355  71.82740 114.44699 123.39107  95.18988 128.92318 124.67930
#> [2829] 122.44236 112.78452 110.13041 104.42941 108.43819  94.48233  80.83732
#> [2836]  96.02312 127.11733 100.57852  68.98949  91.58689  84.70058        NA
#> [2843]  99.96525 119.16547  86.69544  99.49964 143.61958 102.73629  95.60627
#> [2850] 112.25122 121.84025 115.73244  69.28951 101.87189 108.05618 103.63385
#> [2857] 113.60998 151.66784 111.02639 112.87081 138.47606  93.59849  77.20566
#> [2864] 121.33602  84.49641 111.09794  98.62600 105.26697  73.39744  96.03620
#> [2871]  96.38197        NA  90.54799 107.22812 115.93136  90.76849  68.63065
#> [2878]  75.59604  87.72663  85.07567  87.04709  78.97231  90.15908 107.17061
#> [2885]  93.02394  95.20143  93.21813  89.99700 130.57810 115.37659 104.68928
#> [2892]  88.39835 109.88016        NA 113.37751 133.28125 103.54675 117.40363
#> [2899] 116.52549 115.19387 134.25917  89.80796  88.33636 120.39152 108.86024
#> [2906] 111.10999  94.45125  67.00672 104.99094        NA  92.74592 130.05134
#> [2913] 101.82104  93.25730 111.58424  95.73374 100.16999  89.82105 107.90781
#> [2920]  82.48122  79.60684 105.81593  91.15752 115.28435 101.15011        NA
#> [2927] 120.99104  98.40676  97.78318  88.81300 110.08307 104.60101 154.38608
#> [2934] 103.15702  83.42172  66.01947  95.90203 104.11462 113.51574  87.75213
#> [2941]  79.30560  66.06425  81.26630 107.94160  73.49714 106.68626  66.62263
#> [2948] 102.39706 127.11882  91.47758 105.53437 100.46349        NA 130.52706
#> [2955]  96.78742 126.60964  94.30548 136.94105 116.68615  85.12379  98.36577
#> [2962] 107.21394 114.80938 138.38439 101.00855 102.19514 105.96470 122.44937
#> [2969] 120.84356 108.51515 105.10018  90.16884 125.68040  91.69494 114.52728
#> [2976]  81.82861 127.34445 101.89164 113.27872  87.31474 116.49379  96.10415
#> [2983] 100.14702 104.44401 104.97516 108.88994  83.87092  92.98215  99.97636
#> [2990]  81.24271 108.33061 100.81349 105.85727 105.17807 105.76366  91.73068
#> [2997]  99.69245  94.66167 108.41767  92.63342        NA  91.64693 101.70160
#> [3004] 121.42641  95.48389  92.96127  74.12034  99.27988 105.59880 105.26058
#> [3011]        NA 101.71555 103.75203 107.83993 117.78409        NA 103.50437
#> [3018]  96.32741        NA 127.38455 118.37969  94.90177  99.51611        NA
#> [3025]  69.57997  65.99458  81.55889  89.35169 115.92625  99.85068  88.27901
#> [3032] 112.55876 103.44280 109.76347  81.16179  87.28226 113.64760 103.85133
#> [3039] 102.19905        NA  93.66891 102.71730  93.77594  84.88711  97.98131
#> [3046] 121.65572  84.47060  88.17755 119.08629  97.28303  83.57380 126.46446
#> [3053] 110.24537 114.13200 120.36610  86.61578 133.00210        NA 124.97344
#> [3060]  94.38708 150.55091 109.78039  98.35713 106.44163  95.54031        NA
#> [3067] 104.73312  92.29435  87.01589 133.74656  88.95830 117.92075  99.17902
#> [3074] 103.99712 128.45992  94.66730 124.43517  72.56790 103.01083 122.27866
#> [3081] 114.69619 115.11222  85.14453  56.43536  81.44001 149.56714  69.36997
#> [3088]  70.37317 104.08992        NA 102.97098 105.61374  79.39006  76.35066
#> [3095]  93.92220 103.37540 104.00154  89.21541 107.69798 134.43861 101.06532
#> [3102] 112.83415 121.15327 102.87093  85.41530  85.56874 113.74198 132.16915
#> [3109] 104.04807 101.78844 114.09049 144.79477 123.11842  77.45673 111.10628
#> [3116]  94.79047  86.68614  97.91256  66.59719  96.88089  99.31000  99.91169
#> [3123]  91.68684  89.52588  73.20135 110.90477  99.44425 117.97963 126.49695
#> [3130]  99.32024  97.87872  98.04536  85.94370  75.99010 105.25503 105.48335
#> [3137]  82.65481        NA 107.62278 125.99438  72.69358 110.72839  67.86381
#> [3144] 102.24616  97.50902 112.48910  93.80934  95.25180 108.69227 129.90128
#> [3151] 102.61137 117.97364  84.00165  82.12655 115.28914  90.98558 121.43007
#> [3158] 105.80824 103.78158  78.04950 122.04040  74.87177 103.31413 134.82085
#> [3165]  91.71280  83.48088 128.75948 101.42446        NA  82.07126  99.96834
#> [3172]  88.42897  84.40717 147.05392 125.48424  98.38601  88.85027  59.73402
#> [3179] 151.40616 102.59013        NA 109.15048 149.02478  87.19711 144.07494
#> [3186] 120.16093 103.06875  85.50834 110.78064  71.45947 100.27382  81.42361
#> [3193]  89.10960 114.12733  73.39031 101.86087 108.80346 118.24206  94.74655
#> [3200] 116.22270 102.08499  81.82055  83.43677  78.04402  88.98584  86.82979
#> [3207] 132.31632 104.17464 126.75613 114.18304  91.25765 104.63753  98.04227
#> [3214] 152.96094  98.81755 103.21930  92.15451 114.37154  89.79170 138.34373
#> [3221]  93.89319  96.59086  84.37213  67.06667  97.78431 133.64749 151.11687
#> [3228] 105.90173 116.02315 120.51476  79.77432  94.05978 125.62207  85.26982
#> [3235]  99.88914 104.20201 101.37923 115.34974 101.23666 140.58416  81.56064
#> [3242]        NA 109.50642 140.97253 102.10558  75.28914 109.94468  80.12058
#> [3249] 141.73077  84.40061  95.10738  73.16149 132.51380  73.47270 130.93655
#> [3256]  85.83935 112.81693 115.18689  74.12210 118.76769 116.87306  69.77772
#> [3263]  86.86765  91.23787 103.45878 112.23566 111.67880 109.66774 106.96926
#> [3270] 128.75764 119.28980  82.03259  97.82282 139.77222  89.21713 104.03452
#> [3277] 107.03995 106.22589 101.26873 118.22418  91.58290  95.77568 123.46339
#> [3284]  86.92730  92.13656  80.29465 102.72891  91.54427 102.05915  88.94557
#> [3291] 114.30098  88.34672 100.02050 100.65471 107.58998 116.38628  84.56820
#> [3298]  73.97515 117.88154 116.63784 105.13293 122.42326  99.18007  73.06769
#> [3305]  83.56804  96.58232  92.96067 100.14735 109.00791 101.56609 109.12685
#> [3312] 103.60409  90.76348  78.17195 101.25058 108.35830 117.14167  83.97870
#> [3319] 102.65209  90.38504        NA 120.75148 111.47376  81.05197        NA
#> [3326] 116.83046  55.39800 117.13875  78.75607 119.56779  84.40017 120.72403
#> [3333] 102.63719 117.79238 116.38289  96.82403 154.12097 107.20193 134.74513
#> [3340]  80.12545        NA  87.43281 113.09723  93.71942  78.35138 100.37741
#> [3347] 114.80814 109.27697 136.58576  97.99844 109.36557 105.47338 102.09935
#> [3354]  84.59720 125.80947 115.84621  81.88816 126.58897 105.52151 119.76989
#> [3361] 100.11348  88.57029        NA  79.37162 141.07499 103.82288 113.18695
#> [3368] 136.44462 102.33860  82.18938  89.24359        NA 130.37187  91.18344
#> [3375]  89.31767 135.09771 108.55925 135.09482 104.02961 101.45738 122.36591
#> [3382] 127.11527 104.01972  90.60462 125.73082 110.06440  79.50450  94.35962
#> [3389] 127.96560  95.55637 109.94012 102.80316 114.38426  95.41072  90.83146
#> [3396] 116.42672 121.96186 113.21977  89.63267  97.12212 108.70147 113.51258
#> [3403] 108.48103 114.04225 105.03910 128.07731  97.60751 123.13022 119.23357
#> [3410] 120.69198 100.41992 118.08329  91.09615        NA  82.18679  98.61385
#> [3417]  94.07375 106.23787  87.28975 123.83159  89.17761  89.88304 116.90993
#> [3424]  79.60099 153.57694 112.58620  91.95810  88.92122  84.47236  84.05556
#> [3431] 100.22503 102.16883  99.85873  88.43833 105.10660 109.19452  97.24864
#> [3438]  98.53825 131.86360 119.99642 102.49135 105.09698 114.35476 117.80060
#> [3445] 112.10422 118.50496 119.63996  91.80546 126.77410  85.65424        NA
#> [3452] 116.23552 105.75307  90.32062 108.36197  90.34928  92.44760 101.24107
#> [3459] 120.71417 111.23657 132.39479  94.36587 106.25676 131.39871 127.77236
#> [3466] 105.07239 134.31622 127.63342 103.15662 113.91850 115.17813  91.37081
#> [3473] 124.18756  99.29416 117.39452 106.48765        NA 130.96250  84.19303
#> [3480] 112.93341 140.55295 113.58349 100.07731  78.49378  88.31260  97.65220
#> [3487]  99.39208  92.65892  95.45342 111.00992 105.85294 111.12579  96.25577
#> [3494]        NA 100.85954  80.40417 111.49855 147.03408 124.74174 125.00162
#> [3501]  77.56077  87.16312 151.36476  94.15639 101.82416 112.74459 100.78403
#> [3508]  83.30977  75.78516 125.85689        NA 117.41034 109.37199 121.79131
#> [3515]  99.57711  67.17589 109.50676  83.52852  77.40478 100.21845 101.60368
#> [3522]  66.25266  95.54459  91.62957 117.99941 107.69596 111.15091 104.22016
#> [3529]  88.10770  66.73014  95.26508 121.83781  77.15082 103.19277  74.32404
#> [3536] 103.82181 143.18074  98.53710 129.26684 110.03282 118.04621 102.57737
#> [3543] 104.09786 103.61333 126.47433 125.82084  93.37566 116.03461  86.65183
#> [3550] 125.59315 103.14450 121.53041 114.80932  91.79228 107.91343  84.88029
#> [3557]  90.40666  89.80897 104.61652  79.06729  78.19095  99.86568 101.04037
#> [3564] 114.34638 101.58425  56.94059  80.63387  84.81862  78.11127  96.04882
#> [3571]  85.86003 103.48518  98.07238  92.02708 114.02477 111.20562 128.13272
#> [3578]  92.97453 120.97553  96.93657  87.77838  93.86339 117.48021  93.81707
#> [3585] 107.63425 113.35900 111.18742 125.73866 121.13839  83.51386  81.23401
#> [3592]  65.61877  98.78242 105.13664 116.31534  89.27355 121.21548 103.51004
#> [3599]        NA  96.33854 120.76328 107.94670  83.45318 106.68502  97.99816
#> [3606] 118.47178 131.17970        NA  97.25266 103.94992 108.72629  60.79784
#> [3613]  75.25867  83.86011 109.74176  96.53575  77.38883 108.27072        NA
#> [3620]  81.05050 101.60280  99.41720  96.69805 115.28153 141.61014        NA
#> [3627]  99.90698  99.08268 129.93746 104.61920  71.82387  74.07527  98.07306
#> [3634]  89.20965 110.40019 113.24011  72.68010 108.36504  89.81855 139.58302
#> [3641]  97.63460  50.80849  71.29545  94.43073 106.70478  94.99260 128.60525
#> [3648]  84.53487  95.25804  87.66562 121.22734  99.63499 149.67218  98.76202
#> [3655] 129.96943 103.74643 124.36021        NA        NA  91.05552 128.47730
#> [3662] 107.30950  93.10215        NA 128.18690 103.48018  96.83185  92.15587
#> [3669] 107.44630 119.71844 104.59461  92.69617 135.90515 119.64882 116.24849
#> [3676]  98.45496  90.29992 119.05951 140.22358  81.64954  97.46359 109.40464
#> [3683] 115.85938  59.21015  96.17072 118.71853  88.61795 109.43283  93.42442
#> [3690] 119.49987  96.80893  94.20176  63.94520 101.16927 103.29223 125.35324
#> [3697]  73.07509 126.51422 119.22465  86.01306 120.94667  91.34657 127.58170
#> [3704]  94.92981 130.25062 135.61405  78.04711 102.36111 121.21145 112.27543
#> [3711] 103.48925  98.10541  89.30041 101.06586 122.32049 120.08835  94.07367
#> [3718]  64.75347 114.75440 113.28043 100.99591  60.65250 130.71141 106.27882
#> [3725]  73.61821  98.76869  82.44057  98.51861 156.11932  97.10330  76.93554
#> [3732] 100.92091 123.70953  82.81654 107.73891  87.69366        NA 102.57049
#> [3739] 111.08473  95.78498  94.49906 102.83403  75.16502  80.07673  91.80342
#> [3746] 129.58393 107.44467  99.68652 106.66176 131.52500 114.10138  69.57645
#> [3753]  74.77259 118.88603 130.98556 109.85914 131.69319 104.64965  99.96484
#> [3760] 142.46084 121.36070 117.04830 142.42482 101.77788 101.56947 126.86109
#> [3767] 141.06057  95.04049 102.54793 117.12335  66.82641 117.38353 100.07812
#> [3774] 108.32477  93.65935 105.84012  91.96469 101.26281 127.94511 121.35225
#> [3781] 113.02943 114.77231  66.06101 101.51895 100.11878 107.46514  98.19240
#> [3788] 142.09301  96.63460  80.34375 114.64444 126.15649 128.60802 112.28102
#> [3795]  91.83400  88.29544 100.10629  92.88554 131.62088 123.07094  85.75031
#> [3802] 109.23207  94.88870  90.84498  93.16312  95.56644 102.13304  83.75248
#> [3809]  90.61108 108.58566  91.46654 122.33493  99.42630 132.41998 132.78117
#> [3816]  73.90757 130.82880 156.11932 115.90343 112.90506  91.10209 102.92882
#> [3823] 112.42633 105.65504 101.60669  86.16568 101.84428 112.06699 118.32901
#> [3830] 111.28812  97.92133 112.14178 112.19679 113.71858  78.76471  75.73721
#> [3837] 100.61086 102.72400 113.63072 123.41794 112.29610 100.55207  62.99023
#> [3844]  89.08907 128.34980  75.10316        NA  96.42997 126.36364 143.10396
#> [3851]  95.85157  92.44249  87.66238  79.53649  93.80633 109.26248 128.37698
#> [3858] 110.34695  97.42418 121.23882 101.84050  98.46168  94.15634 116.72197
#> [3865] 105.16490 114.08232 102.09025  87.75596 115.84259 115.15956  96.61815
#> [3872] 121.34332 110.02937 140.89450  85.10854 105.81679        NA  81.98802
#> [3879] 115.20350 114.90729  86.41909  75.83175  90.59911 112.02073  81.39721
#> [3886]  94.55745        NA 114.78485 115.61133 131.52324 122.72747 118.88904
#> [3893]  95.10829  71.66612  81.58683 134.51754  94.69288  99.71818 150.05422
#> [3900]  85.36208 105.47713        NA  78.11283  95.30848  75.78084 106.07737
#> [3907]  92.43915  84.35250  97.56549  84.31386  97.94781 114.44628 108.31273
#> [3914] 102.33983 105.28998 112.53219 120.19840 111.46468  68.92859 120.32783
#> [3921]  86.93742 112.44555 113.54955 147.80651  91.16725 119.12227  96.46828
#> [3928]  88.85408  98.97760  82.49134 102.04463 126.25571 119.43158 121.76859
#> [3935] 104.42854  94.70879 108.92350 127.34387 102.23730 127.82833  91.81455
#> [3942]        NA  86.11769  91.79406 103.88267 100.54082 108.53172 106.27484
#> [3949]  75.47172 120.50943 131.51316 109.49381 109.87617 117.79583 103.23942
#> [3956] 118.89085 106.72599  81.09400 112.48335 103.95544 135.87763 102.60540
#> [3963]  96.39302 103.24371 121.43967  83.75791  92.41182  95.92592 119.46730
#> [3970] 109.47459  66.79442  98.48603 102.56792  81.03338  98.15338  90.94396
#> [3977] 117.67134  91.31776 108.27246  85.58059  91.61794 116.31322  93.53823
#> [3984]  98.09188 111.66716  84.62267 110.12182  76.45717  96.88300  89.70734
#> [3991]  78.50915 126.71136 129.07585  73.20348        NA  86.23383  92.35880
#> [3998]        NA 105.83874 117.30097 119.71062  85.16058  92.11620 100.80740
#> [4005] 115.95288  84.21587  83.72864  78.10121  90.47606  81.02392 111.68492
#> [4012]  88.49500 119.02951 111.41247  92.73279 115.00284 124.50126  99.01302
#> [4019]  97.05870  99.30917 112.92501  85.47028  98.10343 126.09843 109.16254
#> [4026] 105.80394  93.27139  72.75953 139.62787  89.17073  80.54986 101.09647
#> [4033]        NA 107.43036  79.67429  88.62061 104.40557  94.91026 111.99022
#> [4040]  92.44407  87.73461        NA  81.96767 108.01709 112.41548  83.22510
#> [4047]  68.43265  77.00571  77.73206  82.79417 108.32232  84.58294  95.97138
#> [4054] 113.14632  64.21573        NA 103.66789  70.09709 135.44159 121.58607
#> [4061]  99.82145 140.54720 128.90643  86.01912        NA  76.23750  88.00970
#> [4068]  89.37556  86.73796  95.06655 103.50127  92.30781  98.15482 134.51852
#> [4075] 108.07388 118.32644  75.56006  94.50226 125.78980        NA  87.65661
#> [4082]  77.63667  97.99957  82.15574 114.78674  98.98158 123.10309  89.92549
#> [4089] 129.58890  82.88296        NA  93.88602 127.63673 113.86251        NA
#> [4096] 105.39922  81.77309 103.21605 118.98158  92.26072        NA 104.09782
#> [4103] 113.50288 100.96712 125.77977 118.47998  95.15550 112.20170 100.69768
#> [4110] 105.35633 109.65963  70.74478 121.16530 110.29490 114.66099 105.68732
#> [4117]  99.49953  95.11391 101.96606 139.30202  96.60142 108.69029  99.22776
#> [4124]  80.55114  99.20488  96.72112  81.83573 115.14132 135.35506  91.42978
#> [4131]        NA 102.15730 104.00859  96.60304 106.62508 120.80180  99.46964
#> [4138]  83.19684 109.03645 116.55311  95.06862 101.66208  99.30697 114.29295
#> [4145]  97.07505 103.72605 110.73079 138.42446 101.08955  73.16837  87.28916
#> [4152]  97.51314  93.02835 141.76325  93.45038  88.69890 118.82198  86.76597
#> [4159] 108.93900 137.39294 116.60359 132.35694  92.57706  99.53116  77.56836
#> [4166]        NA  92.88043 113.69727 104.71006        NA  91.62295 124.24197
#> [4173]  94.46284 104.32443  92.77116  93.57738  91.05293  91.07443        NA
#> [4180]  90.69218  85.89833  88.76489        NA  88.97420 127.23839 119.10226
#> [4187] 134.18466 124.30105 101.67326  88.45767  96.32488 110.38928  86.99683
#> [4194]  80.46743 122.94355 101.23560 118.58970  93.23175 132.60719 112.87660
#> [4201]  91.49588 106.96123  90.31864  98.55323  76.80744  88.84034 128.84678
#> [4208]  88.86738  76.58666  88.28567 149.44579 113.27335 100.77275 121.65146
#> [4215]  67.03969 129.80767  93.77459  85.53127  92.40856 130.92614  72.68711
#> [4222]  89.48707 102.95023 102.49554 102.36131  79.54772 112.75430 131.29114
#> [4229] 108.58857  70.88293 106.50033 121.34966  96.05522  94.79149  84.48531
#> [4236] 101.27907  84.40894 100.35483  89.98260  89.91328  82.47798  82.63364
#> [4243] 123.06684 105.82580 118.94616 111.06931 101.83146 135.48679  99.46511
#> [4250]  68.58877 111.32486  91.19514  90.95705 103.60981  95.97853  91.08264
#> [4257] 102.47083  90.04780  97.61434 110.33878  89.20344  94.48848  73.23116
#> [4264] 104.48700  73.73767  94.27132 108.52765  91.65707  96.03865 127.01862
#> [4271]  82.71180  72.20658 116.06291  75.44421  86.07503 106.48344 100.23460
#> [4278]  86.92888 107.71762 143.42760  62.03382        NA  93.03033 139.34429
#> [4285] 107.01659 117.61955  96.50192 123.87954 119.92701 113.31435 113.19898
#> [4292] 119.90833  69.80516  87.31722  85.17203 105.86567 140.04342 139.02863
#> [4299]  97.89270  87.39024 112.08209  85.26482 143.54806 155.81189 130.53565
#> [4306]  80.91873  80.56678  85.63642 101.64851 125.36324 106.83964 100.84729
#> [4313] 116.32227 106.63242  75.46333  81.04244 100.32530 133.61492 112.49000
#> [4320] 125.52991 116.68293  76.35093 124.64920  92.10817 110.94964  93.15704
#> [4327]  95.63793  98.37136  92.95598  72.66921  90.33524 101.82716 113.54016
#> [4334] 102.51488 135.54290  80.53427 116.85788  80.37707 114.67669  83.17996
#> [4341]  89.96941  89.34390  91.06010  67.41105  91.04868 134.42893  83.70470
#> [4348] 102.16165 111.91678  75.18333 145.79586 113.92883 109.89556        NA
#> [4355] 127.42700 113.82127 134.26772  94.98811 114.00394 133.85516 117.12830
#> [4362] 115.75440  90.72054  95.01097  97.94184 113.17038  99.64314  97.59306
#> [4369] 104.07762 142.17487 102.83811 109.66167  79.01423  79.37186  87.87358
#> [4376] 123.65926  81.94108  76.74082  80.28829 143.18629  95.77493  73.29896
#> [4383] 117.03640  91.34827 103.61848 116.34491 111.04633 117.40872  87.96507
#> [4390] 114.17102  99.94776 111.90707  95.28089 118.70722 118.42164 111.97829
#> [4397]  89.85899  93.39832 100.09364 110.06767 106.03296  90.06741 124.85853
#> [4404] 116.88761 105.40621  93.30190 111.74609  84.26113 102.99222  58.96093
#> [4411] 105.96499 100.80566  89.96737  89.23849  89.52275 116.25939 107.82514
#> [4418]  92.53411 114.57573 109.15086 113.04176 111.80949 115.77704 112.83377
#> [4425]  83.47754  97.31830 119.49189 118.50881 139.03202  89.49087  71.40488
#> [4432] 111.68749 104.10706 112.43726  93.26594  93.24696 110.42388 135.34162
#> [4439] 125.46328 108.27092  83.04188 109.04359 103.46716  79.09586 101.68991
#> [4446]  86.44726 102.61473 148.22268  86.26580 112.39280  91.62882  88.65460
#> [4453]  99.26960 106.68452 121.26133  95.00700 132.34570  93.23510 113.47603
#> [4460]  98.04631 115.84413  57.77654  94.59543  86.74046  97.34384 117.33182
#> [4467]  99.69141  91.60432  92.82277 115.94391 108.81819 127.02630 116.48781
#> [4474] 120.19797 108.67838 114.67771 108.52482  99.93686 156.11932 105.87798
#> [4481] 113.41031  95.94625 139.00624  88.03825  98.64246        NA  98.49621
#> [4488] 113.05204  94.32566 111.90114 115.84281 121.82250 144.27380  98.34798
#> [4495]  93.08917  91.04316 117.59979  95.55865  81.50188  94.73665  71.93240
#> [4502] 110.05042        NA  82.99127  93.73088  94.02659 112.05853  98.34527
#> [4509]  90.62678 140.47608  99.83789        NA  77.27460  82.27287  90.46273
#> [4516] 105.35766  72.43714 133.53620 113.24704  62.92348 120.99707  86.83537
#> [4523] 118.45424 109.86415 129.34252 103.24181 125.09423        NA  76.53308
#> [4530] 105.53199 109.69992  99.41518  95.32681 103.05730 122.75130 117.82625
#> [4537] 105.46893 107.44208 113.19272  98.92471  94.19762 104.08396  89.28204
#> [4544]  98.27039 101.82329 106.71684 105.09285  67.90658 111.07796  96.72455
#> [4551]  65.67897        NA 111.75265 120.18356 123.42744  97.30558  74.47298
#> [4558] 136.99919  81.44144  88.11105  84.10000 128.14547 104.03495  84.74085
#> [4565]  98.96928 104.96646 140.20822 109.09581 100.73673 106.77906 117.95522
#> [4572]  85.95598 102.67402 109.48128 112.58099  98.94319  92.97245 127.66684
#> [4579] 121.22080  91.33405 110.47474 102.07256 111.02614 124.41379 130.38367
#> [4586] 115.39561  90.79966  89.39972 110.03605 116.30821  92.75861 131.33713
#> [4593]  64.93535 113.19281 112.96955        NA  99.47067  86.61926 105.65400
#> [4600] 112.69177 105.32577 107.10007 100.65929 101.75379        NA        NA
#> [4607] 109.71580 106.37319 130.99941  85.30391 112.05519  76.31054  79.14479
#> [4614] 115.99498 110.55388  93.40940  82.24704        NA 108.60088  99.87256
#> [4621] 103.55116  79.11574 130.84572 120.32583        NA  99.67104  89.14890
#> [4628]  77.65489 108.71118 119.88905 149.08465 100.52447 124.10560  78.24761
#> [4635] 115.05429 130.97541  88.35891 125.20356  98.00008 100.94018 133.00368
#> [4642] 122.83475  88.43547 121.11382 107.15763 129.64542  95.46090 119.64615
#> [4649]        NA 135.01409 103.00706 107.08680 104.20170 101.08518 112.99926
#> [4656] 101.38578  81.73822 118.12043 120.61390  80.37557 108.15004 133.78634
#> [4663]  92.83672  96.90677  76.85624 112.43471  98.60968 131.91148 134.76571
#> [4670]  97.78536  98.50090 109.37147        NA 138.51817 128.78439  68.43324
#> [4677]  82.92148 124.69364 109.10235  88.15406 116.83341 102.98207  99.53311
#> [4684] 123.45263  94.81939  99.26011  83.22405  92.80933        NA 105.84595
#> [4691]  88.70541  95.80608 147.60688  65.27416 130.70225  99.02732 127.50559
#> [4698] 103.63945 125.82559  89.13104  86.63387  94.89105 107.90086 113.56329
#> [4705]        NA        NA        NA  77.14036 132.60901  79.08299  75.35484
#> [4712]  92.44554  99.17820  87.93566  91.41259  78.35827  83.64300 106.59775
#> [4719]  98.80397 110.52985 132.48631  98.45999 103.99550 138.33839  93.15327
#> [4726] 100.29292 116.08963  97.14672  87.30811 124.95039  92.27693  93.08749
#> [4733]  71.20865 100.51909  74.88666 128.33117 115.15241 115.36160  81.40905
#> [4740]  92.96084 116.64495  95.57341        NA 144.30328 100.48700  98.76096
#> [4747]  85.51842  79.93950  99.73590 108.23107  97.46928        NA 123.75043
#> [4754]  84.93888  94.75774  76.51885  77.06462  71.56516  93.38056  80.25712
#> [4761] 125.34671 113.15372  65.47597 110.12225 112.94309 117.24581        NA
#> [4768] 105.32930 105.44843  87.38285  95.28276 126.11150 100.74478  98.21451
#> [4775] 107.94462 113.31007  95.64928 120.06386 100.50336  98.34996  94.81563
#> [4782] 110.34383  63.75114 115.72456  99.72184  87.54991  92.56199  96.97001
#> [4789] 105.40717  99.40382 106.53898  91.81409  90.01876 145.17715  58.98819
#> [4796] 133.45365 108.43753  87.52302 103.37834 108.02134  95.52201  91.76117
#> [4803]  97.60954  90.42905 100.67166  82.78712 103.10171 103.88756  84.57005
#> [4810]  69.15095 100.36170 149.55872 115.40036  99.75839 124.17054 118.25758
#> [4817] 117.32375 115.30221  87.37881 104.21563 104.23538 132.89353 106.77036
#> [4824] 107.90755  81.73038 130.75804 105.74168  92.17668  73.19464 106.70845
#> [4831] 104.14742 101.42816 119.53426  97.73236  88.98683 131.33996  85.99529
#> [4838]  75.81940  90.62757 101.53924 112.93163 128.50055 106.82083 111.15069
#> [4845] 106.14997  93.28999  74.68711 108.95380 137.40403 113.56338  92.97683
#> [4852]  95.46528  58.79528  81.04308 108.56046  99.29878 109.46275 117.76990
#> [4859] 108.78428  99.59168  77.63904  87.22594 120.58785 120.18065 118.46264
#> [4866] 102.15798 119.69481 153.60518 134.55506  87.87020 101.32607  94.40437
#> [4873]  65.06705 132.34647 107.80524  63.11078 107.47843 129.92524 100.28326
#> [4880] 112.78122 116.62816 130.85799 129.63147  78.73503  81.75735 125.78381
#> [4887] 117.46420 104.96620 103.69841  84.89960  93.00182  98.11861 133.00308
#> [4894]  95.07406 104.06231  91.18422  91.03861 107.81537 127.71394  99.70869
#> [4901]  95.47614  90.48390 106.12803  84.03679  78.68849 107.79692  95.69611
#> [4908] 104.23555 156.11932  90.29213  90.90127  77.57228  90.80795  93.22094
#> [4915] 108.58288  72.49457  83.42553 119.23394 110.33314  91.43836 122.47413
#> [4922]  87.87023  94.55722        NA 116.21738 128.08320  91.16688 107.66351
#> [4929] 116.78613 100.73225 126.69054  79.94986 116.16952  92.99823  98.79995
#> [4936] 100.49874 130.03652 130.72129 132.22988 121.46700  93.87933 103.56240
#> [4943]  94.63712 114.00421 105.79924        NA 112.42900 119.90781  83.12623
#> [4950]  84.94425 136.10615  92.25194 101.91107  93.64825 144.87292 116.79199
#> [4957]  86.48469  92.26783 101.45355  96.72760  91.07416  78.43076  71.72591
#> [4964]  88.59788 118.64055 108.09569 116.39752 111.41827  86.35791 107.03181
#> [4971]  85.11046 101.52113  83.77892  89.39695 107.66816  55.20972  96.78332
#> [4978] 115.00098  84.41108  94.34485 124.81015  70.31361  97.65948 104.49745
#> [4985]  98.91200 121.34653  92.03801 129.51919  97.82457  89.61659 112.12695
#> [4992]  78.91838 110.45574 116.39745 101.20386  93.07578  75.78792  98.13211
#> [4999]  79.75552 108.62884

if (FALSE) {
psydat %>%
  dplyr::select(c(iq, anxT)) %>%
  map(winsorZ)
psydat %>% mutate_at(c("iq", "anxT"), list(~ winsorZ(.)))
psydat %>% mutate_if(is.double, list(~ winsorZ(.)))
}